+++ R/Confirmatory-methods.R
@@ -4,32 +4,8 @@ setMethod(
     signature = signature(x = 'ConfirmatoryClass'),
     definition = function(x)
     {
-        cat("\nCall:\n", paste(deparse(x@Call), sep = "\n", collapse = "\n"),
-            "\n\n", sep = "")
-        cat("Full-information item factor analysis with ", x@nfact, " factors \n", sep="")
-        EMquad <- ''
-        if(x@method == 'EM') EMquad <- c(' with ', x@quadpts, ' quadrature')
-        if(x@converge == 1)
-            cat("Converged in ", x@iter, " iterations", EMquad, ". \n", sep = "")
-        else
-            cat("Estimation stopped after ", x@iter, " iterations", EMquad, ". \n", sep="")
-        if(!is.nan(x@condnum))
-            cat("Condition number of information matrix = ", x@condnum, 
-                '\nSecond-order test: model ', if(!x@secondordertest) 
-                    'is a possible saddle point (non-maximum)' else 
-                    'is a possible local maximum', '\n', sep = "")
-        if(length(x@logLik) > 0){
-            cat("Log-likelihood = ", x@logLik, ifelse(length(x@SElogLik) > 0,
-                                                               paste(', SE = ', round(x@SElogLik,3)),
-                                                               ''), "\n",sep='')
-            cat("AIC = ", x@AIC, "; AICc = ", x@AICc, "\n", sep='')
-            cat("BIC = ", x@BIC, "; SABIC = ", x@SABIC, "\n", sep='')
-            if(!is.nan(x@p)){
-                cat("G2 (", x@df,") = ", round(x@G2,2), ", p = ", round(x@p,4), sep='')
-                cat("\nRMSEA = ", round(x@RMSEA,3), ", CFI = ", round(x@CFI,3),
-                    ", TLI = ", round(x@TLI,3), sep='')
-            }
-        }
+        class(x) <- 'ExploratoryClass'
+        print(x)
     }
 )
 
@@ -71,199 +47,49 @@ setMethod(
 setMethod(
     f = "coef",
     signature = 'ConfirmatoryClass',
-    definition = function(object, CI = .95, digits = 3, IRTpars = FALSE, rawug = FALSE, ...)
+    definition = function(object, ...)
     {
-        if(CI >= 1 || CI <= 0)
-            stop('CI must be between 0 and 1')
-        z <- abs(qnorm((1 - CI)/2))
-        SEnames <- paste0('CI_', c((1 - CI)/2*100, ((1 - CI)/2 + CI)*100))
-        K <- object@K
-        J <- length(K)
-        nLambdas <- ncol(object@F)
-        allPars <- list()
-        if(IRTpars){
-            if(object@nfact > 1L)
-                stop('traditional parameterization is only available for unidimensional models')
-            for(i in 1:(J+1))
-                allPars[[i]] <- round(mirt2traditional(object@pars[[i]]), digits)
-        } else {
-            if(length(object@pars[[1]]@SEpar) > 0){
-                for(i in 1:(J+1)){
-                    allPars[[i]] <- round(matrix(c(object@pars[[i]]@par,
-                                                   object@pars[[i]]@par - z*object@pars[[i]]@SEpar,
-                                                   object@pars[[i]]@par + z*object@pars[[i]]@SEpar),
-                                             3, byrow = TRUE), digits)
-                    rownames(allPars[[i]]) <- c('par', SEnames)
-                    colnames(allPars[[i]]) <- names(object@pars[[i]]@est)
-                }
-            } else {
-                for(i in 1:(J+1)){
-                    allPars[[i]] <- matrix(round(object@pars[[i]]@par, digits), 1L)
-                    colnames(allPars[[i]]) <- names(object@pars[[i]]@est)
-                    rownames(allPars[[i]]) <- 'par'
-                }
-            }
-        }
-        if(!rawug){
-            allPars <- lapply(allPars, function(x, digits){
-                x[ , colnames(x) %in% c('g', 'u')] <- round(antilogit(x[ , colnames(x) %in% c('g', 'u')]), digits)
-                x
-            },  digits=digits)
-        }
-        names(allPars) <- c(colnames(object@data), 'GroupPars')
-        return(allPars)
+        class(object) <- 'ExploratoryClass'
+        coef(object,  ...)
     }
 )
 
 setMethod(
     f = "residuals",
     signature = signature(object = 'ConfirmatoryClass'),
-    definition = function(object, restype = 'LD', digits = 3, df.p = FALSE, full.scores = FALSE,
-                          printvalue = NULL, tables = FALSE, verbose = TRUE, ...)
+    definition = function(object, ...)
     {
-        K <- object@K
-        data <- object@data
-        N <- nrow(data)
-        J <- ncol(data)
-        nfact <- ncol(object@F) - length(attr(object@pars, 'prodlist'))
-        itemloc <- object@itemloc
-        res <- matrix(0,J,J)
-        diag(res) <- NA
-        colnames(res) <- rownames(res) <- colnames(data)
-        theta <- seq(-4,4, length.out = round(20/nfact))
-        Theta <- thetaComb(theta,nfact)
-        ThetaShort <- Theta
-        if(length(object@prodlist) > 0) Theta <- prodterms(Theta, object@prodlist)
-        gpars <- ExtractGroupPars(object@pars[[length(object@pars)]])
-        prior <- mvtnorm::dmvnorm(ThetaShort,gpars$gmeans,gpars$gcov)
-        prior <- prior/sum(prior)
-        df <- (object@K - 1) %o% (object@K - 1)
-        diag(df) <- NA
-        colnames(df) <- rownames(df) <- colnames(res)
-        itemnames <- colnames(data)
-        listtabs <- list()
-        if(restype == 'LD'){
-            for(i in 1:J){
-                for(j in 1:J){
-                    if(i < j){
-                        P1 <- ProbTrace(x=object@pars[[i]], Theta=Theta)
-                        P2 <- ProbTrace(x=object@pars[[j]], Theta=Theta)
-                        tab <- table(data[,i],data[,j])
-                        Etab <- matrix(0,K[i],K[j])
-                        for(k in 1:K[i])
-                            for(m in 1:K[j])
-                                Etab[k,m] <- N * sum(P1[,k] * P2[,m] * prior)
-                        s <- gamma.cor(tab) - gamma.cor(Etab)
-                        if(s == 0) s <- 1
-                        res[j,i] <- sum(((tab - Etab)^2)/Etab) * sign(s)
-                        res[i,j] <- sign(res[j,i]) * sqrt( abs(res[j,i]) / (N*min(c(K[i],K[j]) - 1L)))
-                        df[i,j] <- pchisq(abs(res[j,i]), df=df[j,i], lower.tail=FALSE)
-                        if(tables){
-                            tmp <- paste0(itemnames[i], '_', itemnames[j])
-                            listtabs[[tmp]] <- list(Obs=tab, Exp=Etab, std_res=(tab-Etab)/sqrt(Etab))
-                        }
-                    }
-                }
-            }
-            if(tables) return(listtabs)
-            if(df.p){
-                cat("Degrees of freedom (lower triangle) and p-values:\n\n")
-                print(round(df, digits))
-                cat("\n")
-            }
-            if(verbose) cat("LD matrix (lower triangle) and standardized values:\n\n")
-            res <- round(res,digits)
-            return(res)
-        }
-        if(restype == 'exp'){
-            r <- object@tabdata[ ,ncol(object@tabdata)]
-            res <- round((r - object@Pl * nrow(object@data)) /
-                sqrt(object@Pl * nrow(object@data)),digits)
-            expected <- round(N * object@Pl,digits)
-            tabdata <- object@tabdata
-            rownames(tabdata) <- NULL
-            ISNA <- is.na(rowSums(tabdata))
-            expected[ISNA] <- res[ISNA] <- NA
-            tabdata <- data.frame(tabdata,expected,res)
-            colnames(tabdata) <- c(colnames(object@tabdata),"exp","res")
-            if(full.scores){
-                tabdata[, 'exp'] <- object@Pl / r * N
-                tabdata2 <- object@tabdatalong
-                tabdata2 <- tabdata2[,-ncol(tabdata2)]
-                stabdata2 <- apply(tabdata2, 1, paste, sep='', collapse = '/')
-                sfulldata <- apply(object@fulldata, 1, paste, sep='', collapse = '/')
-                scoremat <- tabdata[match(sfulldata, stabdata2), 'exp', drop = FALSE]
-                res <- (1-scoremat) / sqrt(scoremat)
-                colnames(res) <- 'res'
-                ret <- cbind(object@data, scoremat, res)
-                ret[is.na(rowSums(ret)), c('exp', 'res')] <- NA
-                rownames(ret) <- NULL
-                return(ret)
-            } else {
-                tabdata <- tabdata[do.call(order, as.data.frame(tabdata[,1:J])),]
-                if(!is.null(printvalue)){
-                    if(!is.numeric(printvalue)) stop('printvalue is not a number.')
-                    tabdata <- tabdata[abs(tabdata[ ,ncol(tabdata)]) > printvalue, ]
-                }
-                return(tabdata)
-            }
-        }
+        class(object) <- 'ExploratoryClass'
+        residuals(object, ...)
     }
 )
 
 setMethod(
     f = "anova",
     signature = signature(object = 'ConfirmatoryClass'),
-    definition = function(object, object2, verbose = TRUE)
+    definition = function(object, object2, ...)
     {
-        nitems <- length(object@K)
-        if(length(object@df) == 0 || length(object2@df) == 0)
-            stop('Use \'logLik\' to obtain likelihood values')
-        df <- object@df - object2@df
-        if(df < 0){
-            tmp <- object
-            object <- object2
-            object2 <- tmp
-        }
-        X2 <- round(2*object2@logLik - 2*object@logLik, 3)
-        if(verbose){
-            cat('\nModel 1: ')
-            print(object@Call)
-            cat('Model 2: ')
-            print(object2@Call)
-            cat('\n')
-        }
-        ret <- data.frame(AIC = c(object@AIC, object2@AIC),
-                          AICc = c(object@AICc, object2@AICc),
-                          SABIC = c(object@SABIC, object2@SABIC),
-                          BIC = c(object@BIC, object2@BIC),
-                          logLik = c(object@logLik, object2@logLik),
-                          X2 = c('', X2),
-                          df = c('', abs(df)),
-                          p = c('', round(1 - pchisq(X2,abs(df)),3)))
-        return(ret)
+        class(object) <- 'ExploratoryClass'
+        anova(object, object2, ...)
     }
 )
 
 setMethod(
     f = "fitted",
     signature = signature(object = 'ConfirmatoryClass'),
-    definition = function(object, digits = 3, ...){
-        tabdata <- object@tabdata
-        N <- nrow(object@data)
-        expected <- round(N * object@Pl,digits)
-        return(cbind(tabdata,expected))
+    definition = function(object, ...){
+        class(object) <- 'ExploratoryClass'
+        fitted(object, ...)
     }
 )
 
 setMethod(
     f = "plot",
     signature = signature(x = 'ConfirmatoryClass', y = 'missing'),
-    definition = function(x, y, type = 'info', npts = 50, theta_angle = 45,
-                          rot = list(xaxis = -70, yaxis = 30, zaxis = 10), ...)
+    definition = function(x, y, ...)
     {
         class(x) <- 'ExploratoryClass'
-        plot(x, type=type, npts=npts, theta_angle=theta_angle, rot=rot, ...)
+        plot(x, ...)
 
     }
 )