+++ R/MultipleGroup-methods.R
@@ -4,34 +4,8 @@ setMethod(
     signature = signature(x = 'MultipleGroupClass'),
     definition = function(x)
     {
-        cat("\nCall:\n", paste(deparse(x@Call), sep = "\n", collapse = "\n"),
-            "\n\n", sep = "")
-        cat("Full-information item factor analysis with ", x@nfact, " factors \n", sep="")
-        EMquad <- ''
-        if(x@method == 'EM') EMquad <- c(' with ', x@quadpts, ' quadrature')
-        if(x@converge == 1)
-            cat("Converged in ", x@iter, " iterations", EMquad, ". \n", sep = "")
-        else
-            cat("Estimation stopped after ", x@iter, " iterations", EMquad, ". \n", sep="")
-        if(!is.nan(x@condnum))
-            cat("Condition number of information matrix = ", x@condnum, 
-                '\nSecond-order test: model ', if(!x@secondordertest) 
-                    'is a possible saddle point (non-maximum)' else 
-                        'is a possible local maximum', '\n', sep = "")
-        if(length(x@logLik) > 0){
-            cat("Log-likelihood = ", x@logLik, ifelse(length(x@SElogLik) > 0,
-                                                      paste(', SE = ', round(x@SElogLik,3)),
-                                                      ''), "\n",sep='')
-            cat("AIC = ", x@AIC, "; AICc = ", x@AICc, "\n", sep='')
-            cat("BIC = ", x@BIC, "; SABIC = ", x@SABIC, "\n", sep='')
-            if(!is.nan(x@p)){
-                cat("G2 (", x@df,") = ", round(x@G2,2), ", p = ", round(x@p,4), sep='')
-                cat("\nRMSEA = ", round(x@RMSEA,3), ", CFI = ", round(x@CFI,3),
-                    ", TLI = ", round(x@TLI,3), '\n\n', sep='')
-                for(g in 1:length(x@cmods))
-                    cat(as.character(x@groupNames[g]), " group: G2 = ", round(x@cmods[[g]]@G2,2), '\n', sep='')
-            }
-        }
+        class(x) <- 'ExploratoryClass'
+        print(x)
     }
 )
 
@@ -85,34 +59,10 @@ setMethod(
 setMethod(
     f = "anova",
     signature = signature(object = 'MultipleGroupClass'),
-    definition = function(object, object2, verbose = TRUE)
+    definition = function(object, object2, ...)
     {
-        nitems <- length(object@K)
-        if(length(object@df) == 0 || length(object2@df) == 0)
-            stop('Use \'logLik\' to obtain likelihood values')
-        df <- object@df - object2@df
-        if(df < 0){
-            tmp <- object
-            object <- object2
-            object2 <- tmp
-        }
-        X2 <- round(2*object2@logLik - 2*object@logLik, 3)
-        if(verbose){
-            cat('\nModel 1: ')
-            print(object@Call)
-            cat('Model 2: ')
-            print(object2@Call)
-            cat('\n')
-        }
-        ret <- cbind(AIC = c(object@AIC, object2@AIC),
-                          AICc = c(object@AICc, object2@AICc),
-                          SABIC = c(object@SABIC, object2@SABIC),
-                          BIC = c(object@BIC, object2@BIC),
-                          logLik = c(object@logLik, object2@logLik),
-                          X2 = c(NA, X2),
-                          df = c(NA, abs(df)),
-                          p = c(NA, round(1 - pchisq(X2,abs(df)),3)))
-        ret
+        class(object) <- 'ExploratoryClass'
+        anova(object, object2, ...)
     }
 )